requirements.txt
altgraph==0.17
appdirs==1.4.4
buildozer==1.2.0
certifi==2020.12.5
chardet==4.0.0
cx-Freeze==6.5.2
distlib==0.3.1
docutils==0.16
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==3.4.0
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
numpy==1.19.5
opencv-python==4.5.1.48
pefile==2019.4.18
pexpect==4.8.0
Pillow==8.1.0
ptyprocess==0.7.0
Pygments==2.7.4
pyinstaller @ https://github.com/pyinstaller/pyinstaller/tarball/develop
pyinstaller-hooks-contrib==2020.11
pylibdmtx==0.1.9
pypiwin32==223
pywin32==300
pywin32-ctypes==0.2.0
pyzbar==0.1.8
requests==2.25.1
sh==1.14.1
six==1.15.0
urllib3==1.26.2
virtualenv==20.3.1
zipp==3.4.0
